Convert drawText to use int for position as we shouldn't draw at non-integer positions

+++ nibble/NibbleGUI.cpp
@@ -84,8 +84,8 @@ void NibbleGUI::DrawTextAligned(const char* text, Rectangle bounds, Rectangle pa
 	DrawTextEx(font, text, { (float)x, (float)y }, font.baseSize, font_spacing, colour);
 }
 
-void NibbleGUI::drawText(std::string text_str, float x, float y, Color colour) {
-	DrawTextEx(font, text_str.c_str(), { x, y }, font.baseSize, font_spacing, colour);
+void NibbleGUI::drawText(std::string text_str, int x, int y, Color colour) {
+	DrawTextEx(font, text_str.c_str(), { (float)x, (float)y }, font.baseSize, font_spacing, colour);
 }
 
 void NibbleGUI::setGUIColourPalette(GuiPalette gui_palette) { palette = gui_palette; }
